05-usage-data/python-analysis/requirements.txt
# FullStory Data Analysis Dependencies

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests for API calls
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=7.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Date/time handling
python-dateutil>=2.8.0

# Progress bars for long operations
tqdm>=4.65.0

# Optional: Statistical analysis
# scipy>=1.11.0
# statsmodels>=0.14.0
